+++ src/main/java/org/openlmis/requisition/domain/requisition/Requisition.java
@@ -20,12 +20,13 @@ import static java.util.function.Function.identity;
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toMap;
 import static org.apache.commons.lang3.BooleanUtils.isNotTrue;
-import static org.apache.commons.lang3.BooleanUtils.isTrue;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 import static org.openlmis.requisition.domain.requisition.RequisitionLineItem.ADJUSTED_CONSUMPTION;
 import static org.openlmis.requisition.domain.requisition.RequisitionLineItem.AVERAGE_CONSUMPTION;
 import static org.openlmis.requisition.domain.requisition.RequisitionLineItem.CALCULATED_ORDER_QUANTITY;
 import static org.openlmis.requisition.i18n.MessageKeys.ERROR_FIELD_MUST_HAVE_VALUES;
+import static org.openlmis.requisition.i18n.MessageKeys.ERROR_LINE_ITEM_ADDED;
+import static org.openlmis.requisition.i18n.MessageKeys.ERROR_LINE_ITEM_REMOVED;
 import static org.openlmis.requisition.i18n.MessageKeys.ERROR_MUST_BE_INITIATED_TO_BE_SUBMMITED;
 import static org.openlmis.requisition.i18n.MessageKeys.ERROR_MUST_BE_SUBMITTED_TO_BE_AUTHORIZED;
 
@@ -786,10 +787,8 @@ public class Requisition extends BaseTimestampedEntity {
           .orElse(null);
 
       if (null == existing) {
-        if (isTrue(emergency) || item.isNonFullSupply()) {
-          item.setRequisition(this);
-          updatedList.add(item);
-        }
+        item.setRequisition(this);
+        updatedList.add(item);
       } else {
         existing.setRequisition(this);
         existing.updateFrom(item);
@@ -798,17 +797,25 @@ public class Requisition extends BaseTimestampedEntity {
     }
 
     if (isNotTrue(emergency)) {
-      // is there a full supply line that is not in update list
-      // it should be added. Those lines should not be removed
-      // during update. Only non full supply lines can be
-      // added/updated/removed.
-      List<UUID> updatedIds = updatedList.stream().map(BaseEntity::getId).collect(toList());
-
-      requisitionLineItems
+      List<UUID> currentIds = updatedList
           .stream()
           .filter(line -> !line.isNonFullSupply())
-          .filter(line -> !updatedIds.contains(line.getId()))
-          .forEach(updatedList::add);
+          .map(BaseEntity::getId)
+          .collect(toList());
+
+      List<UUID> existingIds = requisitionLineItems
+          .stream()
+          .filter(line -> !line.isNonFullSupply())
+          .map(BaseEntity::getId)
+          .collect(toList());
+
+      if (currentIds.stream().anyMatch(id -> !existingIds.contains(id))) {
+        throw new ValidationMessageException(ERROR_LINE_ITEM_ADDED);
+      }
+
+      if (existingIds.stream().anyMatch(id -> !currentIds.contains(id))) {
+        throw new ValidationMessageException(ERROR_LINE_ITEM_REMOVED);
+      }
     }
 
     requisitionLineItems.clear();